requirements.txt
pyzipper
piexif
Pillow
colorama